image_captioning/requirements.txt
easyocr==1.6.2
nltk==3.7
numpy==1.23.5
Pillow==9.5.0
streamlit==1.21.0
torch==2.0.0
transformers==4.28.1